requirements.txt
streamlit>=1.24.0
Pillow>=9.0.0
tiktoken>=0.5.0
groq>=0.4.0
